+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Unreleased
 
+* Add `SentryAssetBundle` for automatic spans for asset loading (#685)
+
 # 6.3.0-beta.1
 
 * Enha: Replace flutter default root name '/' with 'root' (#678)
 * Fix: Use 'navigation' instead of 'ui.load' for auto transaction operation (#675)
-* Add `SentryAssetBundle` for automatic spans for asset loading (#685)
 * Fix: Use correct data/extras type in tracer (#693)
 * Fix: Do not throw when Throwable type is not supported for associating errors to a transaction (#692)