+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 
-* Bump: Sentry-Android to 5.5.2 and Sentry-Cocoa to 7.8.0 (#)
+* Bump: Sentry-Android to 5.5.2 and Sentry-Cocoa to 7.8.0 (#696)
 
 # 6.3.0-beta.1